requirements.txt
langchain==0.0.189
openai==0.27.6
python-dotenv==1.0.0
Jinja2==3.1.2
streamlit==1.23.1

